Fixed uit64_t functionality

+++ include/BigInt.hpp
@@ -9,7 +9,7 @@
 #define BIG_INT_HPP
 
 #include <iostream>
-#include <stdint.h>
+#include <cstdint>
 #include <vector>
 
 class BigInt {
@@ -20,12 +20,12 @@ class BigInt {
         // Constructors:
         BigInt();
         BigInt(const BigInt&);
-        BigInt(const long long&);
+        BigInt(const int64_t&);
         BigInt(const std::string&);
 
         // Assignment operators:
         BigInt& operator=(const BigInt&);
-        BigInt& operator=(const long long&);
+        BigInt& operator=(const int64_t&);
         BigInt& operator=(const std::string&);
 
         // Unary arithmetic operators:
@@ -38,11 +38,11 @@ class BigInt {
         BigInt operator*(const BigInt&) const;
         BigInt operator/(const BigInt&) const;
         BigInt operator%(const BigInt&) const;
-        BigInt operator+(const long long&) const;
-        BigInt operator-(const long long&) const;
-        BigInt operator*(const long long&) const;
-        BigInt operator/(const long long&) const;
-        BigInt operator%(const long long&) const;
+        BigInt operator+(const int64_t&) const;
+        BigInt operator-(const int64_t&) const;
+        BigInt operator*(const int64_t&) const;
+        BigInt operator/(const int64_t&) const;
+        BigInt operator%(const int64_t&) const;
         BigInt operator+(const std::string&) const;
         BigInt operator-(const std::string&) const;
         BigInt operator*(const std::string&) const;
@@ -55,11 +55,11 @@ class BigInt {
         BigInt& operator*=(const BigInt&);
         BigInt& operator/=(const BigInt&);
         BigInt& operator%=(const BigInt&);
-        BigInt& operator+=(const long long&);
-        BigInt& operator-=(const long long&);
-        BigInt& operator*=(const long long&);
-        BigInt& operator/=(const long long&);
-        BigInt& operator%=(const long long&);
+        BigInt& operator+=(const int64_t&);
+        BigInt& operator-=(const int64_t&);
+        BigInt& operator*=(const int64_t&);
+        BigInt& operator/=(const int64_t&);
+        BigInt& operator%=(const int64_t&);
         BigInt& operator+=(const std::string&);
         BigInt& operator-=(const std::string&);
         BigInt& operator*=(const std::string&);
@@ -79,12 +79,12 @@ class BigInt {
         bool operator>=(const BigInt&) const;
         bool operator==(const BigInt&) const;
         bool operator!=(const BigInt&) const;
-        bool operator<(const long long&) const;
-        bool operator>(const long long&) const;
-        bool operator<=(const long long&) const;
-        bool operator>=(const long long&) const;
-        bool operator==(const long long&) const;
-        bool operator!=(const long long&) const;
+        bool operator<(const int64_t&) const;
+        bool operator>(const int64_t&) const;
+        bool operator<=(const int64_t&) const;
+        bool operator>=(const int64_t&) const;
+        bool operator==(const int64_t&) const;
+        bool operator!=(const int64_t&) const;
         bool operator<(const std::string&) const;
         bool operator>(const std::string&) const;
         bool operator<=(const std::string&) const;
